pyproject.toml
[tool.poetry]
name = "denzonium-bot-python-V2"
version = "0.1.0"
description = ""
authors = ["Denzven Vadakkan <denzvenvadakkan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
wikipedia = "^1.4.0"
youtube-dl = "^2021.4.7"
bs4 = "^0.0.1"
discord = "^1.0.1"
flask = "^1.1.2"
"alexflipnote.py" = "^2.3.0"
discord-pretty-help = "^1.3.0"
"akinator.py" = "^4.1.0"
lavalink = "^3.1.4"
dbdmusic = "^0.0.1"
prsaw = "^0.3.1"
pymongo = "^3.11.3"
dnspython3 = "^1.15.0"
python-chess = "^1.999"
english-words = "^1.0.3"
jishaku = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
